% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_base.R
\name{climex}
\alias{climex}
\title{The Climex app initialization}
\usage{
climex(list.data.sources = NULL, data.frame.positions = NULL,
  resource.directory = NULL)
}
\arguments{
\item{resource.directory}{Folder all the resources of the app,
like JavaScript scripts, images etc., will be copied in. If
NULL, the content will be copied in the directory contained in
the global \emph{climex.path} variable pointing at
\emph{~/R/climex/}. Note that the folder won't be removed
afterwards. Default: NULL.}
}
\value{
starts a shiny app
}
\description{
Shiny app combining most of the tools used in extreme
  value analysis using the generalized extreme value (GEV) or
  generalized Pareto (GP) fitting via maximum likelihood.
}
\details{
This app needs its own css file. In order to assure its
  correct behavior, the folder \emph{resource.directory} will be
  generated to store all the necessary configuration scripts.
}
\seealso{
Other top-level: \code{\link{climex.server}},
  \code{\link{climex.ui}}
}
\author{
Philipp Mueller
}
\concept{top-level}
