% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_base.R
\name{climex.server}
\alias{climex.server}
\title{The \code{climex} app server}
\usage{
climex.server(input, output, session)
}
\arguments{
\item{input}{Namespace input. For more details check out
\url{http://shiny.rstudio.com/articles/modules.html}}

\item{output}{Namespace output.}

\item{session}{Namespace session.}
}
\value{
Function acting as the shiny server.
}
\description{
Server-side part of the \pkg{climexUI} \pkg{shiny}
  application.
}
\details{
All data handled by the server will be stored in the
  \emph{input.RData} file along with all other JavaScript, CSS
  etc. files in a folder, which path is defined in the option
  \code{climex.path}. The \code{\link{climex}} should be used as
  an auxiliary function to generate this folder and to stuff it
  with all required files.
}
\seealso{
Other top-level: \code{\link{climex.ui}},
  \code{\link{climex}}
}
\author{
Philipp Mueller
}
\concept{top-level}
