% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{convert.data.frame.to.xts}
\alias{convert.data.frame.to.xts}
\title{Convert time series}
\usage{
convert.data.frame.to.xts(input.df)
}
\arguments{
\item{input.df}{\code{data.frame},  which has  to  contain  two
columns called  \code{value} and \code{date}. The  former has to
be of class \code{numeric} and the latter of class \code{Date}.}
}
\value{
Either \code{NULL} if the conversion failed or an object
  of class \pkg{xts}.
}
\description{
Converts a time series provided as \code{data.frame}
  into an object of class \pkg{xts}.
}
