% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climex_leaflet.R
\name{data.chosen}
\alias{data.chosen}
\title{Choosing a data set in the \code{climex} app}
\usage{
data.chosen(selectDataBase, sliderYears, climex.environment)
}
\arguments{
\item{selectDataBase}{Character (select) input to determine the
data source. It is either of one of the names of the provided
list in the \code{list.data.sources} argument of the
\code{\link{climex}} function or \emph{Artificial data}. In case
of the latter choice, the function \code{\link{data.selection}}
will provide a \emph{reactive} object containing random numbers
drawn from the distribution specified using
\code{radioEvdStatistics}. Default = a random element of
the provided input.}

\item{sliderYears}{Numerical (slider) input to determine the
minimal length (in years) of the time series to be
displayed. Minimal value is 0 and maximal is 155, the default
value is 65 and the step width is 1.}

\item{climex.environment}{Environment containing the global
variables used within the \code{climex} app. Namely the last
values displayed in the table and the lists containing the
station data.}
}
\value{
Reactive \code{list}  containing  a  \code{list} of  all
  selected stations and their positions.
}
\description{
This functions extracts all stations containing more
  than a specified number of years of data.
}
\details{
It uses the current database and returns all stations,
  which are at least as long as \code{sliderYears}.
}
\seealso{
Other leaflet: \code{\link{leafletClimexUI}},
  \code{\link{leafletClimex}}
}
\author{
Philipp Mueller
}
\concept{leaflet}
