% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climex_fitting.R
\name{data.fitting}
\alias{data.fitting}
\title{Reactive fitting in the Climex app}
\usage{
data.fitting(reactive.extreme, reactive.rows, fit.interactive,
  radioEvdStatistics, buttonMinMax, sliderThreshold, selectDataBase)
}
\arguments{
\item{reactive.extreme}{Reactive value returning a list containing
three elements: 1. the blocked time series, 2. the
deseasonalized time series, and 3. the pure time series.}

\item{reactive.rows}{Reactive value holding a logical vector
indicating which values of the time series provided by
\code{\link{data.extremes}} to use after clicking and brushing.}

\item{fit.interactive}{Function used to perform the actual GEV/GP
fit. \code{\link{fit.interactive}}}

\item{radioEvdStatistics}{Character (radio) input determining
whether the GEV or GP distribution shall be fitted to the
data. Choices: c( "GEV", "GP" ), default = "GEV".}

\item{buttonMinMax}{Character (radio) input determining whether
the GEV/GP distribution shall be fitted to the smallest or biggest
values. Choices: c( "Max", "Min ), default = "Max".}

\item{sliderThreshold}{Numerical (slider) input determining the
threshold used within the GP fit and the extraction of the extreme
events. Boundaries: minimal and maximal value of the
  deseasonalized time series (rounded). Default: 0.8* the upper
  end point.}

\item{selectDataBase}{Character (select) input to determine the
data source. It is either of one of the names of the provided
list in the \code{list.data.sources} argument of the
\code{\link{climex}} function or \emph{Artificial data}. In case
of the latter choice, the function \code{\link{data.selection}}
will provide a \emph{reactive} object containing random numbers
drawn from the distribution specified using
\code{radioEvdStatistics}. Default = a random element of
the provided input.}
}
\value{
Reactive  value  holding   a  fitted  object  of  class
  \code{climex.fit.gev} or \code{climex.fit.gpd}, depending on the
  choice in \code{radioEvdStatistics}.
}
\description{
A reactive value performing the GEV/GP fit.
}
\details{
The fit is performed using the
  \code{\link{fit.interactive}} function. Using the reactive.rows
  input specific points of the time series can be removed.
}
\seealso{
Other climex-fitting: \code{\link{color.table}},
  \code{\link{fit.interactive}},
  \code{\link{generalFitStatisticsTable}},
  \code{\link{generalFitStatistics}}
}
\author{
Philipp Mueller
}
\concept{climex-fitting}
