% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_data.R
\name{data.frame.positions.fallback}
\alias{data.frame.positions.fallback}
\title{Longitude, latitude, and altitude information for the station data
provided in the \code{list.data.sources.fallback} data within this
package. It is a data.frame of one row and four columns.}
\format{Class \code{data.frame}}
\description{
Longitude, latitude, and altitude information for the station data
provided in the \code{list.data.sources.fallback} data within this
package. It is a data.frame of one row and four columns.
}
\seealso{
Other fallbackData: \code{\link{list.data.sources.fallback}}
}
\concept{fallbackData}
