% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climex_sidebar.R
\name{data.selection}
\alias{data.selection}
\title{Data selection in the \code{climex} app}
\usage{
data.selection(reactive.chosen, selectDataSource, selectDataBase,
  sliderThreshold, radioEvdStatistics, sliderArtificialDataLocation,
  sliderArtificialDataScale, sliderArtificialDataShape, reactive.redrawing,
  sliderSeriesLength)
}
\arguments{
\item{reactive.chosen}{Reactive value containing a list of the
list of all provided stations and a \code{data.frame} containing
the meta data.}

\item{selectDataSource}{Menu output in the sidebar. It is a
\code{character string} describing the selected station's name.}

\item{selectDataBase}{Character (select) input to determine the
data source. It is either of one of the names of the provided
list in the \code{list.data.sources} argument of the
\code{\link{climex}} function or \emph{Artificial data}. In case
of the latter choice, the function \code{\link{data.selection}}
will provide a \emph{reactive} object containing random numbers
drawn from the distribution specified using
\code{radioEvdStatistics}. Default = a random element of
the provided input.}

\item{sliderThreshold}{Numerical (slider) input determining the
threshold used within the GP fit and the extraction of the
extreme events. Boundaries: minimal and maximal value of the
deseasonalized time series (rounded). Default: 0.8* the upper
end point.}

\item{radioEvdStatistics}{Character (radio) input determining
whether the GEV or GP distribution shall be fitted to the
data. Choices: c( "GEV", "GP" ), default = "GEV".}

\item{sliderArtificialDataLocation}{Numerical (slider) input
providing the location parameter to generate an artificial time
series. If \code{radioEvdStatistics} equals \code{"GP"} or
\code{selectDataBase} does not equal \code{"Artificial data"},
this argument will be \code{NULL}.}

\item{sliderArtificialDataScale}{Numerical (slider) input
providing the scale parameter to generate an artificial time
series. If \code{selectDataBase} does not equal
\code{"Artificial data"}, this argument will be \code{NULL}.}

\item{sliderArtificialDataShape}{Numerical (slider) input
providing the shape parameter to generate an artificial time
series. If \code{selectDataBase} does not equal\code{"Artificial
data"}, this argument will be \code{NULL}.}

\item{reactive.redrawing}{Reactive event triggering whenever the
user presses the \emph{Draw} button displayed when
\code{"Artificial data"}' is chosen in \code{selectDataBase}. If
not chosen, this button will not be defined.}

\item{sliderSeriesLength}{Numerical slider determining the length
of an artificial time series. This one will only be present when
\code{selectDataBase} is set to \code{"Artificial data"}.}
}
\value{
Reactive value containing a \pkg{xts} class time series.
}
\description{
Reactive value selecting an individual time series.
}
\details{
According to the choice in \code{selectDataBase} an
  artificial time series will be sampled or one from a database
  will be selected. For the latter one the reactive value
  \code{\link{data.chosen}} will be constulted. The length of
  the generated time series is determined by the
  \code{sliderSeriesLength}.
}
\seealso{
Other sidebar: \code{\link{sidebarCleaningInput}},
  \code{\link{sidebarCleaning}},
  \code{\link{sidebarDataBaseInput}},
  \code{\link{sidebarDataBase}},
  \code{\link{sidebarDataSourceInput}},
  \code{\link{sidebarDataSource}},
  \code{\link{sidebarDataTypeInput}},
  \code{\link{sidebarDataType}},
  \code{\link{sidebarImprintInput}},
  \code{\link{sidebarImprint}},
  \code{\link{sidebarLoadingGifOutput}},
  \code{\link{sidebarLoadingGif}},
  \code{\link{sidebarLoadingInput}},
  \code{\link{sidebarLoading}},
  \code{\link{sidebarSeriesLengthInput}},
  \code{\link{sidebarSeriesLength}}
}
\author{
Philipp Mueller
}
\concept{sidebar}
