% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climex_preprocessing.R
\name{deseasonalize.interactive}
\alias{deseasonalize.interactive}
\title{Removing seasonality in the \code{climex} app}
\usage{
deseasonalize.interactive(x.xts, selectDeseasonalize, selectDataBase)
}
\arguments{
\item{x.xts}{Time  series of  class \pkg{xts},  which has  to be
cleaned.}

\item{selectDeseasonalize}{\code{Character}   (select)  input
determining, which  deseasonalization method  should be  used to
remove  the  short-range  correlations from  the  provided  time
series.       \code{\link{deseasonalizeSelectionInput}}.      If
\code{NULL} \code{\link{anomalies}} will be used.}

\item{selectDataBase}{Character  (select) input to  determine the
data source.  It  is either of one of the  names of the provided
list   in   the   \code{list.data.sources}   argument   of   the
\code{\link{climex}} function or \emph{Artificial data}. In case
of the latter  choice, the function \code{\link{data.selection}}
will provide a \emph{reactive}  object containing random numbers
drawn     from     the      distribution     specified     using
\code{radioEvdStatistics}. Default =  a random element of
the provided input.}
}
\value{
Time series of class \pkg{xts}.
}
\description{
Function for removing the seasonality of a given time
  series within the \code{climex} app.
}
\seealso{
Other preprocessing: \code{\link{cleaning.interactive}},
  \code{\link{data.extremes}},
  \code{\link{deseasonalizeSelectionInput}},
  \code{\link{deseasonalizeSelection}},
  \code{\link{extremes.interactive}},
  \code{\link{generalButtonMinMaxInput}},
  \code{\link{generalButtonMinMax}},
  \code{\link{generalExtremeExtractionInput}},
  \code{\link{generalExtremeExtraction}}
}
\author{
Philipp Mueller
}
\concept{preprocessing}
