% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climex_preprocessing.R
\name{deseasonalizeSelection}
\alias{deseasonalizeSelection}
\title{Removing the seasonality in the \code{climex} app}
\usage{
deseasonalizeSelection(selectDataBase)
}
\arguments{
\item{selectDataBase}{Character  (select) input to  determine the
data source.  It is either of  one of the names  of the provided
list   in   the   \code{list.data.sources}   argument   of   the
\code{\link{climex}} function or \emph{Artificial data}. In case
of the latter  choice, the function \code{\link{data.selection}}
will provide a \emph{reactive}  object containing random numbers
drawn     from     the      distribution     specified     using
\code{radioEvdStatistics}. Default =  a random element of
the provided input.}
}
\value{
selectInput
}
\description{
Removing the seasonality.
}
\details{
Not  a real \pkg{shiny} module, since I  have to use this
  select input outside its namespace.
}
\seealso{
Other preprocessing: \code{\link{cleaning.interactive}},
  \code{\link{data.extremes}},
  \code{\link{deseasonalize.interactive}},
  \code{\link{deseasonalizeSelectionInput}},
  \code{\link{extremes.interactive}},
  \code{\link{generalButtonMinMaxInput}},
  \code{\link{generalButtonMinMax}},
  \code{\link{generalExtremeExtractionInput}},
  \code{\link{generalExtremeExtraction}}
}
\author{
Philipp Mueller
}
\concept{preprocessing}
