% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climex_plotting.R
\name{function.get.y.label}
\alias{function.get.y.label}
\title{y plot labels in the Climex app}
\usage{
function.get.y.label(selectDataBase, selectDataType)
}
\arguments{
\item{selectDataBase}{Character (select) input to determine the
data source. In the default installation there are two
options: c( "Input", "Artificial data" ). The first one
uses the database loaded at the initialization of the climex app
(see the \code{\link{climex}}). The second one allows the user 
to produce random numbers distributed according to the GEV or GP
distribution. Determined by menuSelectDataBase. Default =
"Input".}

\item{selectDataType}{Character (select) input to determine which
set of measurements should be used. This choice is important if
the input of the \code{\link{climex}} function was not just a
list of different station data, but a list of such lists. This
additional layer of lists can e.g. represent different types of
measurement data like precipitation and temperature. Their names
are derived from the names of the input list (see \code{link{
menuSelectDataType}}).}
}
\value{
Character
}
\description{
Determines the propoper label of the y label of the
  plots in the General tab of the Climex app.
}
\details{
It's more or less just the input$selectDataType being
  considered. For artificial data the label will be
  "temperature".
}
\seealso{
Other climex-plot: \code{\link{generalFitPlotOutput}},
  \code{\link{generalFitPlot}},
  \code{\link{generalTimeSeriesPlotOutput}},
  \code{\link{generalTimeSeriesPlot}}
}
\author{
Philipp Mueller
}
