% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climex_preprocessing.R
\name{generalButtonMinMax}
\alias{generalButtonMinMax}
\title{Minimum or maximum extremes in the Climex app}
\usage{
generalButtonMinMax(radioEvdStatistics, selectDataType)
}
\arguments{
\item{radioEvdStatistics}{Character (radio) input determining
whether the GEV or GP distribution shall be fitted to the
data. Choices: c( "GEV", "GP" ), default = "GEV".}

\item{selectDataType}{Character (select) input to determine which
set of measurements should be used. This choice is important if
the input of the \code{\link{climex}} function was not just a
list of different station data, but a list of such lists. This
additional layer of lists can e.g. represent different types of
measurement data like precipitation and temperature. Their names
are derived from the names of the input list (see \code{link{
menuSelectDataType}}).}
}
\value{
menuItemOutput
}
\description{
Whether to determine the minimal or maximal extremes.
}
\details{
Not a real shiny module, since I have to use this select
  input outside its namespace. Only when input$radioEvdStatistics
  is set of "GEV" the minimal extremes scan be used.
}
\seealso{
Other preprocessing: \code{\link{cleaning.interactive}},
  \code{\link{data.extremes}},
  \code{\link{deseasonalize.interactive}},
  \code{\link{deseasonalizeSelectionInput}},
  \code{\link{deseasonalizeSelection}},
  \code{\link{extremes.interactive}},
  \code{\link{generalButtonMinMaxInput}},
  \code{\link{generalExtremeExtractionInput}},
  \code{\link{generalExtremeExtraction}}
}
\author{
Philipp Mueller
}
