% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climex_leaflet.R
\name{leafletClimexUI}
\alias{leafletClimexUI}
\title{Leaflet interface of the \code{climex} app}
\usage{
leafletClimexUI(id)
}
\arguments{
\item{id}{Namespace prefix}
}
\value{
\code{\link[shiny]{tagList}}
}
\description{
Leaflet  interface to  select stations  or visualize
  spacial information.
}
\details{
This   function   provides  the   user  interface   to
  \code{\link{leafletClimex}}.         It       consists        of
  \code{\link[leaflet]{leafletOutput}}            and            two
  \code{\link[shiny]{absolutePanel}} containing  the return level
  and time series length sliders as well as  the table displaying
  all the stations information.
}
\seealso{
Other leaflet: \code{\link{data.chosen}},
  \code{\link{leafletClimex}}
}
\author{
Philipp Mueller
}
\concept{leaflet}
