% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_data.R
\name{list.data.sources.fallback}
\alias{list.data.sources.fallback}
\title{Daily  maximum  temperatures of  the  Potsdam  station in  Germany
provided by the DWD. If you intend to use this data please look at
the  \href{ftp://ftp-cdc.dwd.de/pub/CDC/Terms_of_use.pdf}{terms of
use}. The data is stored in the \pkg{xts} format.}
\format{Class \code{list(list(\pkg{xts}))}}
\source{
\url{http://www.dwd.de/}
}
\description{
The series is the same one as provided by the \pkg{climex} package
but formatted in such a way the \pkg{climexUI} functions can
handle it properly.
}
\details{
The object is a \emph{list of length 1}, which is named "daily
max. temp.". This \emph{element is again a list of length 1},
named "Potsdam", which now contains the actual time series of
class \pkg{xts}.
}
\seealso{
Other fallbackData: \code{\link{data.frame.positions.fallback}}
}
\concept{fallbackData}
