% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climex_sidebar.R
\name{sidebarCleaning}
\alias{sidebarCleaning}
\title{Data cleaning in the \code{climex} app}
\usage{
sidebarCleaning(radioEvdStatistics, selectDataBase)
}
\arguments{
\item{radioEvdStatistics}{Character (radio) input determining
whether the GEV or GP distribution shall be fitted to the
data. Choices: c( "GEV", "GP" ), default = "GEV".}

\item{selectDataBase}{Character (select) input to determine the
data source. It is either of one of the names of the provided
list in the \code{list.data.sources} argument of the
\code{\link{climex}} function or \emph{Artificial data}. In case
of the latter choice, the function \code{\link{data.selection}}
will provide a \emph{reactive} object containing random numbers
drawn from the distribution specified using
\code{radioEvdStatistics}. Default = a random element of
the provided input.}
}
\value{
\code{\link[shinydashboard]{renderMenu}}
}
\description{
Toggling the cleaning of the time series.
}
\details{
A checkbox input. If toggled and the GEV distribution was
  chosen via \code{radioEvdStatistics}, all incomplete years will
  be removed from the time series. If on the other hand the GP
  distribution was picked, all cluster will be removed and only
  their highest point will remain.
}
\seealso{
Other sidebar: \code{\link{data.selection}},
  \code{\link{sidebarCleaningInput}},
  \code{\link{sidebarDataBaseInput}},
  \code{\link{sidebarDataBase}},
  \code{\link{sidebarDataSourceInput}},
  \code{\link{sidebarDataSource}},
  \code{\link{sidebarDataTypeInput}},
  \code{\link{sidebarDataType}},
  \code{\link{sidebarImprintInput}},
  \code{\link{sidebarImprint}},
  \code{\link{sidebarLoadingGifOutput}},
  \code{\link{sidebarLoadingGif}},
  \code{\link{sidebarLoadingInput}},
  \code{\link{sidebarLoading}},
  \code{\link{sidebarSeriesLengthInput}},
  \code{\link{sidebarSeriesLength}}
}
\author{
Philipp Mueller
}
\concept{sidebar}
