% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climex_sidebar.R
\name{sidebarLoading}
\alias{sidebarLoading}
\title{Sidebar menu selection in the \code{climex} app}
\usage{
sidebarLoading(session, selectDataBase, radioEvdStatistics)
}
\arguments{
\item{session}{Namespace session.  For details  check out  \url{
http://shiny.rstudio.com/articles/modules.html}}

\item{selectDataBase}{Character  (select) input to  determine the
data source.  It is either of  one of the names  of the provided
list   in   the   \code{list.data.sources}   argument   of   the
\code{\link{climex}} function or \emph{Artificial data}. In case
of the latter  choice, the function \code{\link{data.selection}}
will provide a \emph{reactive}  object containing random numbers
drawn     from     the      distribution     specified     using
\code{radioEvdStatistics}. Default =  a random element of
the provided input.}

\item{radioEvdStatistics}{Character (radio)  input  determining
whether  the GEV  or  GP  distribution shall  be  fitted to  the
data. Choices: c( "GEV", "GP" ), default = "GEV".}
}
\value{
\code{\link[shinydashboard]{renderMenu}}
}
\description{
Sidebar menu selection.
}
\details{
Provides the fourth  selection menu in the  sidebar.  If
  \code{selectDataBase},                provided                by
  \code{\link{sidebarDataBase}},         was        set         to
  \code{"Artificial  data"}, the  menu will  be a  numerical input
  slider to provide the shape parameter for the GEV or \code{NULL}
  for the  GP distribution. Else it  will be a file  input for the
  user to load additional station data.
}
\seealso{
Other sidebar: \code{\link{data.selection}},
  \code{\link{sidebarCleaningInput}},
  \code{\link{sidebarCleaning}},
  \code{\link{sidebarDataBaseInput}},
  \code{\link{sidebarDataBase}},
  \code{\link{sidebarDataSourceInput}},
  \code{\link{sidebarDataSource}},
  \code{\link{sidebarDataTypeInput}},
  \code{\link{sidebarDataType}},
  \code{\link{sidebarImprintInput}},
  \code{\link{sidebarImprint}},
  \code{\link{sidebarLoadingGifOutput}},
  \code{\link{sidebarLoadingGif}},
  \code{\link{sidebarLoadingInput}},
  \code{\link{sidebarSeriesLengthInput}},
  \code{\link{sidebarSeriesLength}}
}
\author{
Philipp Mueller
}
\concept{sidebar}
