% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climex_sidebar.R
\name{sidebarLoadingGifOutput}
\alias{sidebarLoadingGifOutput}
\title{Gif in the \code{climex} app}
\usage{
sidebarLoadingGifOutput(id)
}
\arguments{
\item{id}{Namespace prefix}
}
\value{
\code{\link[shiny]{div}}
}
\description{
Showing a Gif in the sidebar while the app is
  processing/on hold
}
\details{
Rendering of \code{\link{sidebarLoadingGif}}.
}
\seealso{
Other sidebar: \code{\link{data.selection}},
  \code{\link{sidebarCleaningInput}},
  \code{\link{sidebarCleaning}},
  \code{\link{sidebarDataBaseInput}},
  \code{\link{sidebarDataBase}},
  \code{\link{sidebarDataSourceInput}},
  \code{\link{sidebarDataSource}},
  \code{\link{sidebarDataTypeInput}},
  \code{\link{sidebarDataType}},
  \code{\link{sidebarImprintInput}},
  \code{\link{sidebarImprint}},
  \code{\link{sidebarLoadingGif}},
  \code{\link{sidebarLoadingInput}},
  \code{\link{sidebarLoading}},
  \code{\link{sidebarSeriesLengthInput}},
  \code{\link{sidebarSeriesLength}}
}
\author{
Philipp Mueller
}
\concept{sidebar}
