% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climex_sidebar.R
\name{sidebarSeriesLength}
\alias{sidebarSeriesLength}
\title{Series length in the \code{climex} app}
\usage{
sidebarSeriesLength(selectDataBase)
}
\arguments{
\item{selectDataBase}{Character (select) input to determine the
data source. It is either of one of the names of the provided
list in the \code{list.data.sources} argument of the
\code{\link{climex}} function or \emph{Artificial data}. In case
of the latter choice, the function \code{\link{data.selection}}
will provide a \emph{reactive} object containing random numbers
drawn from the distribution specified using
\code{radioEvdStatistics}. Default = a random element of
the provided input.}
}
\value{
\code{\link[shinydashboard]{menuItemOutput}}
}
\description{
Slider to determine the length of an artificial time
  series.
}
\details{
Numerical slider which will only be displayed when
\code{"Artificial data"} is selected in \code{selectDataBase}.
}
\seealso{
Other sidebar: \code{\link{data.selection}},
  \code{\link{sidebarCleaningInput}},
  \code{\link{sidebarCleaning}},
  \code{\link{sidebarDataBaseInput}},
  \code{\link{sidebarDataBase}},
  \code{\link{sidebarDataSourceInput}},
  \code{\link{sidebarDataSource}},
  \code{\link{sidebarDataTypeInput}},
  \code{\link{sidebarDataType}},
  \code{\link{sidebarImprintInput}},
  \code{\link{sidebarImprint}},
  \code{\link{sidebarLoadingGifOutput}},
  \code{\link{sidebarLoadingGif}},
  \code{\link{sidebarLoadingInput}},
  \code{\link{sidebarLoading}},
  \code{\link{sidebarSeriesLengthInput}}
}
\author{
Philipp Mueller
}
\concept{sidebar}
