% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{check.input}
\alias{check.input}
\title{Check whether the input fulfills our requirements.}
\usage{
check.input(list.data.sources, data.frame.positions, silent = FALSE)
}
\arguments{
\item{list.data.sources}{A named list of \pkg{xts}-class time
series.}

\item{data.frame.positions}{A data.frame containing at least the
named columns \emph{name}, \emph{longitude}, and
\emph{latitude}.}
}
\value{
A list containing the result of the check (TRUE if the
  input has the right format) and a version of the input stripped
  of any list elements and data.frame columns, which just appear
  in one of the input arguments.
  \itemize{
    \item{ check.result: Boolean }
    \item{ list.data.sources: Same format as input }
    \item{ data.frame.positions: Same format as input }
 }
}
\description{
The time series have to be provided as a named list
  of class \pkg{xts} objects and the position data has to be a
  data.frame containing at least the named columns \emph{name},
  \emph{longitude}, and \emph{latitude}.
}
\details{
All named elements of `list.data.sources`, which do not
  appear in `data.frame.positions$name` will be discarded and vice
  versa.
}
\author{
Philipp Mueller
}
