% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climex_preprocessing.R
\name{generalExtremeExtraction}
\alias{generalExtremeExtraction}
\title{Extracting extreme events in the Climex app}
\usage{
generalExtremeExtraction(radioEvdStatistics, deseasonalize.interactive,
  selectDeseasonalize, buttonMinMax, reactive.selection, selectDataBase)
}
\arguments{
\item{radioEvdStatistics}{Character (radio) input determining
whether the GEV or GP distribution shall be fitted to the
data. Choices: c( "GEV", "GP" ), default = "GEV".}

\item{deseasonalize.interactive}{Function used to remove
seasonality from a given time
series. \code{\link{deseasonalize.interactive}}}

\item{selectDeseasonalize}{Character (select) input determining
which deseasonalization method should be used to remove the
short-range correlations from the provided time series.
\code{\link{deseasonalizeSelectionInput}}}

\item{buttonMinMax}{Character (radio) input determining whether
the GEV/GP distribution shall be fitted to the smallest or
biggest vales. Choices: c( "Max", "Min ), default = "Max".}

\item{reactive.selection}{Reactive value contains the class
\pkg{xts} time series of the selected input or artificially
generated data chosen via the sidebar or the leaflet
map. \code{\link{data.selection}}}

\item{selectDataBase}{Character (select) input to determine the
data source. In the default installation there are two
options: c( "Input", "Artificial data" ). The first one
uses the database loaded at the initialization of the climex app
(see the \code{\link{climex}}). The second one allows the user 
to produce random numbers distributed according to the GEV or GP
distribution. Determined by menuSelectDataBase. Default =
"Input".}
}
\value{
renderMenu
}
\description{
Extract the extreme events from a given time series.
}
\details{
Provides a slider input to determine either the block
  length (in case of the GEV distribution) or the height of the
  threshold (GP).
}
\seealso{
Other preprocessing: \code{\link{cleaning.interactive}},
  \code{\link{data.extremes}},
  \code{\link{deseasonalize.interactive}},
  \code{\link{deseasonalizeSelectionInput}},
  \code{\link{deseasonalizeSelection}},
  \code{\link{extremes.interactive}},
  \code{\link{generalButtonMinMaxInput}},
  \code{\link{generalButtonMinMax}},
  \code{\link{generalExtremeExtractionInput}}
}
\author{
Philipp Mueller
}
