% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climex_plotting.R
\name{generalTimeSeriesPlot}
\alias{generalTimeSeriesPlot}
\title{Time series plot in the Climex app.}
\usage{
generalTimeSeriesPlot(input, output, session, reactive.extreme,
  selectDataBase, selectDataType, radioEvdStatistics, sliderThreshold,
  buttonMinMax)
}
\arguments{
\item{input}{Namespace input. For more details check out
\url{http://shiny.rstudio.com/articles/modules.html}}

\item{output}{Namespace output.}

\item{session}{Namespace session.}

\item{reactive.extreme}{Reactive value returning a list containing
three elements: 1. the blocked time series, 2. the
deseasonalized time series, and 3. the pure time series.}

\item{selectDataBase}{Character (select) input to determine the
data source. In the default installation there are two
options: c( "Input", "Artificial data" ). The first one
uses the database loaded at the initialization of the climex app
(see the \code{\link{climex}}). The second one allows the user 
to produce random numbers distributed according to the GEV or GP
distribution. Determined by menuSelectDataBase. Default =
"Input".}

\item{selectDataType}{Character (select) input to determine which
set of measurements should be used. This choice is important if
the input of the \code{\link{climex}} function was not just a
list of different station data, but a list of such lists. This
additional layer of lists can e.g. represent different types of
measurement data like precipitation and temperature. Their names
are derived from the names of the input list (see \code{link{
menuSelectDataType}}).}

\item{radioEvdStatistics}{Character (radio) input determining
whether the GEV or GP distribution shall be fitted to the
data. Choices: c( "GEV", "GP" ), default = "GEV".}

\item{sliderThreshold}{Numerical (slider) input determining the
threshold used within the GP fit and the extraction of the
extreme events. Boundaries: minimal and maximal value of the
deseasonalized time series (rounded). Default: 0.8* the upper
end point.}

\item{buttonMinMax}{Character (radio) input determining whether
the GEV/GP distribution shall be fitted to the smallest or
biggest vales. Choices: c( "Max", "Min ), default = "Max".}
}
\value{
Reactive value holding a logical vector indicating which
  values of the time series provided by
  \code{\link{data.extremes}} to use after clicking and brushing.
}
\description{
TabBox containing a plot of the pure and
  deseasonalized time series as well as one of all its extreme
  events.
}
\details{
Both the pure time series and the deseasonalized one are
  rendered using the dygraphs package. The extreme events are
  rendered using the \pkg{ggplot2} package. For the later one
  clicking or brushing points enables the user to exclude
  them. This is handled by a reactive value returned by this
  function.
}
\seealso{
Other climex-plot: \code{\link{generalFitPlotOutput}},
  \code{\link{generalFitPlot}},
  \code{\link{generalTimeSeriesPlotOutput}}
}
\author{
Philipp Mueller
}
\concept{climex-plot}
