% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climex_sidebar.R
\name{sidebarSeriesLength}
\alias{sidebarSeriesLength}
\title{Series length in the Climex app}
\usage{
sidebarSeriesLength(selectDataBase)
}
\arguments{
\item{selectDataBase}{Character (select) input to determine the
data source. In the default installation there are two
options: c( "Input", "Artificial data" ). The first one
uses the database loaded at the initialization of the climex app
(see the \code{\link{climex}}). The second one allows the user 
to produce random numbers distributed according to the GEV or GP
distribution. Determined by menuSelectDataBase. Default =
"Input".}
}
\value{
menuItemOutput
}
\description{
Slider to determine the length of an artificial time
  series.
}
\details{
Numerical slider which will only be displayed when
"Artificial data" is selected in input$selectDataBase.
}
\seealso{
Other sidebar: \code{\link{data.selection}},
  \code{\link{file.loading}},
  \code{\link{sidebarCleaningInput}},
  \code{\link{sidebarCleaning}},
  \code{\link{sidebarDataBaseInput}},
  \code{\link{sidebarDataBase}},
  \code{\link{sidebarDataSourceInput}},
  \code{\link{sidebarDataSource}},
  \code{\link{sidebarDataTypeInput}},
  \code{\link{sidebarDataType}},
  \code{\link{sidebarImprintInput}},
  \code{\link{sidebarImprint}},
  \code{\link{sidebarLoadingGifOutput}},
  \code{\link{sidebarLoadingGif}},
  \code{\link{sidebarLoadingInput}},
  \code{\link{sidebarLoading}},
  \code{\link{sidebarSeriesLengthInput}}
}
\author{
Philipp Mueller
}
